refactor(sidecar): commitments-api modules
