+++ bolt-sidecar/src/api/commitments/mod.rs
@@ -1,0 +1,3 @@
+pub mod commitments;
+mod jsonrpc;
+mod spec;
